[model-gateway] move oai header util to router header util
